requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==7.5.0
exceptiongroup==1.2.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.62.1
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.3
openai==1.65.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.14.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
smolagents==1.10.0
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
zipp==3.21.0
